Create sals-shipping
Control Flow Project - Using conditional statements if/elif/else to create a shipping cost calculator based on a table fo package weight costs. 

Stuck in the idea of code working top to bottom - the cost variable takes different values as you move from the top to the bottom of the script.