Create review-password-generator
In this lesson you learned:

A string is a list of characters.
A character can be selected from a string using its index string_name[index]. These indices start at 0.
A ‘slice’ can be selected from a string. These can be between two indices or can be open-ended, selecting all of the string from a point.
Strings can be concatenated to make larger strings.
Preview: Docs Returns the length of an object, which can either be a sequence or collection.
len()
 can be used to determine the number of characters in a string.
Strings can be iterated through using for 
Preview: Docs Loading link description
loops
.
Iterating through strings opens up a huge potential for applications, especially when combined with conditional statements.

In this exercise I make a username and password generator. I used splicing to produce strings from other strings up to certain indices, and looped through a string to generate a password that is made from moving the username indices one to the right.