Create ThreadShed_project
In this project, I put string manipulation into practice. It was a good way to learn how to use loops again, and a refresher on how to strip, split, find and format strings for use in loops and to retrieve relevant information. It was definitely a challenge for me at the time