Create csv_dictionaries_and_delimiters
This is a good example to read csv files with arbitrary delimiters as dictionaries in python. Here, I open a file and read it as a dictionary, with the delimiter '@'. The goal was to create a list (isbn list) and retrieve the isbn from the csv to append to this list.   To do this, I went through each line in the dictionary and retrieved each value associated with the key 'ISBN' and appended it to the list 'isbn_list'.